third_party/WebKit/LayoutTests/scrollbars/auto-scrollbar-fit-content.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<div id="top" style="float:left; display:none;">
  <div id="scroller" style="overflow: auto; height: 100px;">
    <div style="width: 100px; height: 200px; border: 2px solid blue;"></div>
  </div>
</div>

<script>
description("Make sure a parent block that fits content expands when a child adds an auto scrollbar.");

var topDiv = document.getElementById("top");
var scroller = document.getElementById("scroller");
topDiv.style.display = "block";
topDiv.offsetTop;
shouldBe("scroller.clientWidth", "scroller.scrollWidth");
</script>
