third_party/WebKit/LayoutTests/scrollbars/auto-scrollbar-fit-content-expected.txt

Make sure a parent block that fits content expands when a child adds an auto scrollbar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scroller.clientWidth is scroller.scrollWidth
PASS successfullyParsed is true

TEST COMPLETE
